- Removed compatibility code for unsupported Python versions
