+++ src/chameleon/compat.py
@@ -1,18 +1,5 @@
 import sys
 
-def lookup_attr(obj, key):
-    try:
-        return getattr(obj, key)
-    except AttributeError:
-        exc = sys.exc_info()[1]
-        try:
-            get = obj.__getitem__
-        except AttributeError:
-            raise exc
-        try:
-            return get(key)
-        except KeyError:
-            raise exc
 
 def exec_(code, globs=None, locs=None):
     """Execute code in a namespace."""
@@ -30,7 +17,3 @@ def exec_(code, globs=None, locs=None):
 exec_("""def raise_with_traceback(exc, tb):
     raise type(exc), exc, tb
 """)
-
-
-def next(iter):
-    return iter.next()
