+++ src/chameleon/zpt/template.py
@@ -1,8 +1,3 @@
-try:
-    import ast
-except ImportError:
-    from chameleon import ast25 as ast
-
 from functools import partial
 from os.path import dirname
 from hashlib import md5
